Better gyro, might work

+++ sensorenhet/sensorenhet/gyro.h
@@ -12,9 +12,9 @@
 extern uint16_t gyro_data_done;
 
 void rotateDegrees(uint16_t degrees);
-uint16_t updateGyroData();
-uint8_t returnGyroData(uint8_t gyro_data);
-uint8_t returnDegreesRotated();
+int updateGyroData();
+int returnGyroData(int gyro_data);
+int returnDegreesRotated();
 void resetDegreesRotated();
 
 void initGyro();
